requirements.txt
pandas
numpy
scikit-learn
matplotlib
tensorflow
biopython
pyranges
seaborn
pyarrow
fastparquet